feat: Update how we pull entities from github.
Instead of using the processors(GithubDiscoveryProcessor and
GithOrgReaderProcessor) we use providers.

The reasoning can be found here: https://backstage.io/docs/integrations/github/org#using-a-processor-instead-of-a-provider

> An alternative to using the Provider for ingesting organizational
> entities is to use a Processor. This is the old way that's based on
> registering locations with the proper type and target, triggering the
> processor to run.
>
> The drawback of this method is that it will leave orphaned Group/User
> entities whenever they are deleted on your GitHub server, and you cannot
> control the frequency with which they are refreshed, separately from
> other processors.

So since processors are the old way and can leave orphans, we move to
the new way which is via providers.

This also has the added benefit that we can refresh the users and groups
in the org less often since they don't change as quickly which should
also help with some of our rate limiting issues.