feat!: Hide Techdocs from the UI.
The techdocs feature is a default feature in backstage so it's pretty
difficult to remove it fully from the app but it's easy enough to hide
it from the UI.

We're doing this beacuse we're not currently planning on using the
techdocs plugin to display documentation.  We're gonna be using
ReadTheDocs and docs.openedx.org as the root doc site for now.

In the future we may integrate the two better but for now we remove
techdocs to reduces confusion.