docs: Check for updated base images during docker builds.
We build a docker image as a part of deploying backstage.  While we're
doing that, we should check for updates to the image so we don't just
use the last one we pulled.  This will ensure that any updates to the
base image, including security updates, get pulled into our latest
release.