feat: Add packages to be able to pull users from the org.
In order for user entities to get populated automatically from the
github org we need a User entity catalog provide.  This one pulls user
data from github orgs it has access to and creates user entities for
them in the backstage database.

We add the new plugin and install it in the backend system and add
config for it into the app-config.yaml file.